refactor: extract checkIssues.ts into modular architecture
- Split 746-line monolithic file into 6 focused modules
- Maintain 100% backward compatibility through re-exports
- Extract GitHub API operations to github.ts (50 lines)
- Extract text parsing utilities to parsers.ts (148 lines)
- Extract Projects V2 integration to projects.ts (204 lines)
- Extract AI-powered analysis to llm.ts (148 lines)
- Extract business logic validation to validators.ts (201 lines)
- Create shared types interface in types.ts (59 lines)
- Main orchestrator in index.ts (97 lines)
- All modules under 200 lines per spec requirement
- All existing tests pass without modification
