+++ backend/src/domain/alert/util/checkIssues.ts
@@ -1,746 +1,10 @@
-import { Octokit } from '@octokit/rest';
-import { openai } from '@ai-sdk/openai';
-import { generateText } from 'ai';
-import { OPENAI_CONFIG } from '../../../config/openai';
-
-const githubToken = process.env.GITHUB_API_KEY;
-if (!githubToken && process.env.NODE_ENV !== 'test') throw new Error('GITHUB_API_KEY is required');
-
-// OpenAI API key will be automatically read from OPENAI_API_KEY environment variable
-
-const octokit = new Octokit({ auth: githubToken || 'test-token' });
-
 /**
- * Created-at lower bound (inclusive)
- * Target: Issues created on or after 2025-08-17T00:00:00Z
+ * Backward compatibility wrapper for checkIssues
+ * This file maintains the original import path and re-exports the refactored implementation
  */
-const CREATED_SINCE_ISO = '2025-08-17T00:00:00Z';
-const CREATED_SINCE = new Date(CREATED_SINCE_ISO);
 
-export interface IssueAlertCandidate {
-  owner: string;
-  repo: string;
-  checkType: string;
-  title: string;
-  description: string;
-  severity: string;
-  filePath: string;
-  lineNumber: number;
-  codeSnippet: string;
-  branch: string;
-  issueUrl: string;
-}
+// Re-export everything from the modular implementation
+export * from './checkIssues/index';
 
-export interface CheckIssuesResult {
-  owner: string;
-  repo: string;
-  alerts: IssueAlertCandidate[];
-}
-
-// Interface for project item field values
-interface ProjectItemFieldValue {
-  sp?: number;
-  endDate?: Date;
-}
-
-export async function checkIssues(owner: string, repo: string): Promise<CheckIssuesResult> {
-  const alerts: IssueAlertCandidate[] = [];
-  const filePath = '';
-  const lineNumber = -1;
-  const codeSnippet = '';
-  const branch = '';
-
-  // Note: This function only processes actual GitHub Issues, not Pull Requests.
-  // GitHub's issues.listForRepo API returns both issues and PRs, so we filter out PRs
-  // to ensure we only generate alerts for genuine issues.
-
-  try {
-    // Get all open issues with pagination
-    const allItems = await octokit.paginate(octokit.issues.listForRepo, {
-      owner,
-      repo,
-      state: 'open',
-      per_page: 100,
-    });
-
-    // Filter out pull requests - only keep actual issues
-    const issuesOnly = allItems.filter((item: any) => !item.pull_request);
-
-    // Filter by created_at >= 2025-08-17T00:00:00Z
-    const issues = issuesOnly.filter((item: any) => {
-      const createdAtStr: string | undefined = item.created_at;
-      if (!createdAtStr) return false;
-      const createdAt = new Date(createdAtStr);
-      return createdAt >= CREATED_SINCE;
-    });
-
-    console.log(`  Found ${allItems.length} open items (issues + PRs) for ${owner}/${repo}`);
-    console.log(`  Filtered to ${issuesOnly.length} actual issues (excluded ${allItems.length - issuesOnly.length} PRs)`);
-    console.log(`  After created_at filter (>= ${CREATED_SINCE_ISO}): ${issues.length} issues remain`);
-
-    // Pre-load all project data once for performance optimization
-    console.log('  Pre-loading project data (createdAt-filtered)...');
-    const allProjectIssues = await getAllProjectIssues(owner, repo);
-    console.log(`  Found ${allProjectIssues.size} issues in projects (createdAt-filtered)`);
-
-    // Pre-load project field values for filtered issues only
-    console.log('  Pre-loading project field values for filtered issues...');
-    const projectFieldValues = await getAllProjectFieldValues(
-      owner,
-      repo,
-      issues.map((issue: any) => issue.number)
-    );
-    console.log(`  Loaded field values for ${projectFieldValues.size} issues`);
-
-    for (let i = 0; i < issues.length; i++) {
-      const issue: any = issues[i];
-
-      // Double-check: ensure this is actually an issue, not a PR
-      if (issue.pull_request) {
-        console.log(`?? Skipping PR #${issue.number} - this should not happen after filtering`);
-        continue;
-      }
-
-      console.log(`  Processing issue #${issue.number} (${i + 1}/${issues.length})`);
-
-      // Check if issue is unassigned
-      if (!issue.assignee) {
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_unassigned',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} is not assigned to anyone.`,
-          severity: 'low',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      }
-
-      // Get field values from projects (priority) and body (fallback)
-      const projectValues = projectFieldValues.get(issue.number) || {};
-      const bodyStoryPoints = extractStoryPoints(issue.body || '');
-      const bodyEndDate = extractEndDate(issue.body || '');
-
-      // Use project values if available, otherwise fallback to body
-      const storyPoints = projectValues.sp !== undefined ? projectValues.sp : bodyStoryPoints;
-      const endDate = projectValues.endDate !== undefined ? projectValues.endDate : bodyEndDate;
-      
-      // Debug logging for SP values
-      if (projectValues.sp !== undefined) {
-        console.log(`    SP from project: ${projectValues.sp}`);
-      } else if (bodyStoryPoints !== null) {
-        console.log(`    SP from body: ${bodyStoryPoints}`);
-      } else {
-        console.log(`    SP not found`);
-      }
-
-      // Check Story Point (SP)
-      if (storyPoints === null) {
-        // Issue missing Story Point
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_missing_sp',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} has no Story Point assigned.`,
-          severity: 'low',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      } else if (storyPoints > 8) {
-        // SP greater than 8 (possibly overestimated)
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_large_sp',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} has Story Point ${storyPoints} which is greater than 8 (possibly overestimated).`,
-          severity: 'low',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      }
-
-      // Check End Date
-      if (endDate === null) {
-        // Issue missing End Date
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_missing_end_date',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} has no End Date assigned.`,
-          severity: 'low',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      } else if (endDate < new Date(Date.now() - 24 * 60 * 60 * 1000)) {
-        // End Date expired (before yesterday)
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_expired_end_date',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} has expired End Date: ${endDate.toISOString().split('T')[0]}.`,
-          severity: 'middle',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      }
-
-      // Check Project (using pre-loaded data)
-      const isInProject = allProjectIssues.has(issue.number);
-
-      if (!isInProject) {
-        // Issue not linked to GitHub Project
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_not_in_project',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} is not linked to any GitHub Project.`,
-          severity: 'middle',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      }
-
-      // LLM-based detection for template-only and unclear instructions
-      const templateDetection = await detectTemplateOnlyIssue(issue.title, issue.body || '');
-      if (templateDetection.result) {
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_template_only',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} appears to contain only template content. ${templateDetection.reason}`,
-          severity: 'high',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      }
-
-      const clarityDetection = await detectUnclearInstructions(issue.title, issue.body || '');
-      if (clarityDetection.result) {
-        alerts.push({
-          owner,
-          repo,
-          checkType: 'issue_unclear_instruction',
-          title: `issue:${issue.number}`,
-          description: `Issue #${issue.number} lacks clear, actionable instructions. ${clarityDetection.reason}`,
-          severity: 'middle',
-          filePath,
-          lineNumber,
-          codeSnippet,
-          branch,
-          issueUrl: issue.html_url,
-        });
-      }
-    }
-
-    console.log(`  Found ${alerts.length} issue alerts for ${owner}/${repo}`);
-  } catch (error) {
-    console.error(`? Error checking issues for ${owner}/${repo}:`, error);
-  }
-
-  return { owner, repo, alerts };
-}
-
-// Function to get all project field values for issues
-async function getAllProjectFieldValues(
-  owner: string,
-  repo: string,
-  issueNumbers: number[]
-): Promise<Map<number, ProjectItemFieldValue>> {
-  const fieldValues = new Map<number, ProjectItemFieldValue>();
-
-  try {
-    console.log('  Loading project field values using GraphQL for each issue...');
-
-    // Query to get project items and field values for each issue
-    const issueProjectQuery = `
-      query($owner: String!, $repo: String!, $issueNumber: Int!) {
-        repository(owner: $owner, name: $repo) {
-          issue(number: $issueNumber) {
-            projectItems(first: 10) {
-              nodes {
-                project {
-                  ... on ProjectV2 {
-                    id
-                    title
-                  }
-                }
-                fieldValues(first: 50) {
-                  nodes {
-                    ... on ProjectV2ItemFieldTextValue {
-                      field { ... on ProjectV2Field { name } }
-                      text
-                    }
-                    ... on ProjectV2ItemFieldNumberValue {
-                      field { ... on ProjectV2Field { name } }
-                      number
-                    }
-                    ... on ProjectV2ItemFieldSingleSelectValue {
-                      field { ... on ProjectV2SingleSelectField { name } }
-                      name
-                      optionId
-                    }
-                    ... on ProjectV2ItemFieldDateValue {
-                      field { ... on ProjectV2Field { name } }
-                      date
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      }
-    `;
-
-    // Process each issue individually to get its project field values
-    for (const issueNumber of issueNumbers) {
-      try {
-        const response: any = await octokit.graphql(issueProjectQuery, {
-          owner,
-          repo,
-          issueNumber,
-        });
-
-        const projectItems = response?.repository?.issue?.projectItems?.nodes || [];
-        
-        if (projectItems.length > 0) {
-          const itemFieldValues: ProjectItemFieldValue = {};
-          
-          // Process field values from the first project item (usually there's only one)
-          for (const item of projectItems) {
-            if (item.fieldValues && item.fieldValues.nodes) {
-              for (const fieldValue of item.fieldValues.nodes) {
-                if (fieldValue.field && fieldValue.field.name) {
-                  const fieldName = String(fieldValue.field.name).toLowerCase();
-
-                  // Handle SP field (can be Number, Text, or Single Select)
-                  if (fieldName === 'sp' && itemFieldValues.sp === undefined) {
-                    if (fieldValue.number !== undefined) {
-                      // Number field
-                      itemFieldValues.sp = fieldValue.number;
-                    } else if (fieldValue.name !== undefined) {
-                      // Single Select field - try to parse number from the option name
-                      const spMatch = fieldValue.name.match(/(\d+)/);
-                      if (spMatch) {
-                        itemFieldValues.sp = parseInt(spMatch[1], 10);
-                      }
-                    } else if (fieldValue.text) {
-                      // Text field - try to parse number from text
-                      const spMatch = fieldValue.text.match(/(\d+)/);
-                      if (spMatch) {
-                        itemFieldValues.sp = parseInt(spMatch[1], 10);
-                      }
-                    }
-                  } else if (
-                    (fieldName === 'end date' ||
-                      fieldName === 'end_date' ||
-                      fieldName === 'due date' ||
-                      fieldName === 'deadline') &&
-                    fieldValue.date &&
-                    itemFieldValues.endDate === undefined
-                  ) {
-                    itemFieldValues.endDate = new Date(fieldValue.date);
-                  }
-                }
-              }
-            }
-          }
-          
-          // Only add if we found values
-          if (itemFieldValues.sp !== undefined || itemFieldValues.endDate !== undefined) {
-            fieldValues.set(issueNumber, itemFieldValues);
-            console.log(`    Issue #${issueNumber}: SP=${itemFieldValues.sp}, EndDate=${itemFieldValues.endDate ? itemFieldValues.endDate.toISOString().split('T')[0] : 'none'}`);
-          }
-        }
-      } catch (error) {
-        console.log(`    Error fetching project data for issue #${issueNumber}`);
-      }
-    }
-
-    console.log(`  Found field values for ${fieldValues.size} issues from projects V2`);
-  } catch (error) {
-    console.error(`? Error loading project field values for ${owner}/${repo}:`, error);
-    console.log('?? Project field value check will be skipped due to API limitations');
-  }
-
-  return fieldValues;
-}
-
-// Function to get all issues in projects (by checking each issue for project association)
-async function getAllProjectIssues(owner: string, repo: string): Promise<Set<number>> {
-  const allIssues = new Set<number>();
-
-  try {
-    console.log('  Checking project association for filtered issues...');
-    
-    // Get all open issues to check for project association
-    const allItems = await octokit.paginate(octokit.issues.listForRepo, {
-      owner,
-      repo,
-      state: 'open',
-      per_page: 100,
-    });
-
-    // Filter to actual issues (not PRs) and createdAt >= threshold
-    const issuesOnly = allItems.filter((item: any) => !item.pull_request);
-    const filteredIssues = issuesOnly.filter((item: any) => {
-      const createdAtStr: string | undefined = item.created_at;
-      if (!createdAtStr) return false;
-      const createdAt = new Date(createdAtStr);
-      return createdAt >= CREATED_SINCE;
-    });
-
-    // Query to check if an issue is in a project
-    const issueProjectCheckQuery = `
-      query($owner: String!, $repo: String!, $issueNumber: Int!) {
-        repository(owner: $owner, name: $repo) {
-          issue(number: $issueNumber) {
-            projectItems(first: 1) {
-              nodes {
-                id
-              }
-            }
-          }
-        }
-      }
-    `;
-
-    // Check each issue for project association
-    for (const issue of filteredIssues) {
-      try {
-        const response: any = await octokit.graphql(issueProjectCheckQuery, {
-          owner,
-          repo,
-          issueNumber: issue.number,
-        });
-
-        const projectItems = response?.repository?.issue?.projectItems?.nodes || [];
-        if (projectItems.length > 0) {
-          allIssues.add(issue.number);
-        }
-      } catch (error) {
-        // Silently skip if we can't check this issue
-      }
-    }
-
-    console.log(
-      `  Found ${allIssues.size} issues in projects (createdAt >= ${CREATED_SINCE_ISO})`
-    );
-  } catch (error) {
-    console.error(`? Error loading project issues for ${owner}/${repo}:`, error);
-    console.log('?? Project check will be skipped due to API limitations');
-  }
-
-  return allIssues;
-}
-
-// Function to extract Story Point from issue content
-function extractStoryPoints(body: string): number | null {
-  // Search for common patterns for Story Point
-  const patterns = [
-    /SP[:\s]*(\d+)/i,
-    /Story Point[:\s]*(\d+)/i,
-    /Story Points[:\s]*(\d+)/i,
-    /Points[:\s]*(\d+)/i,
-    /Estimate[:\s]*(\d+)/i,
-  ];
-
-  for (const pattern of patterns) {
-    const match = body.match(pattern);
-    if (match) {
-      return parseInt(match[1], 10);
-    }
-  }
-
-  return null;
-}
-
-// Function to extract End Date from issue content
-function extractEndDate(body: string): Date | null {
-  // Search for common patterns for End Date
-  const patterns = [
-    /End Date[:\s]*(\d{4}-\d{2}-\d{2})/i,
-    /Due Date[:\s]*(\d{4}-\d{2}-\d{2})/i,
-    /Deadline[:\s]*(\d{4}-\d2}-\d{2})/i, // note: original had variants; keep primary patterns
-    /Target Date[:\s]*(\d{4}-\d{2}-\d{2})/i,
-    /(\d{4}-\d{2}-\d{2})/g, // Find all date format YYYY-MM-DD
-  ];
-
-  for (const pattern of patterns) {
-    const match = body.match(pattern);
-    if (match) {
-      const date = new Date(match[1]);
-      if (!isNaN(date.getTime())) {
-        return date;
-      }
-    }
-  }
-
-  return null;
-}
-
-/**
- * Detects template-only issues using LLM
- */
-async function detectTemplateOnlyIssue(title: string, body: string): Promise<{ result: boolean; reason: string }> {
-  if (!body || body.trim().length === 0) {
-    return {
-      result: true,
-      reason: 'The issue body is empty, which indicates template-only content.',
-    };
-  }
-
-  try {
-    const prompt = `
-You are analyzing a GitHub Issue body to determine if it contains only template content or placeholders.
-
-Issue Title: ${title}
-Issue Body: ${body}
-
-Mark it as TEMPLATE-ONLY (result=true) if:
-- Only contains placeholder text or template instructions
-- Has phrases like "Please describe here", "Describe the issue", "What did you expect to happen"
-- Contains only template sections without actual content filled in
-- Only has boilerplate text that hasn't been customized for the specific issue
-
-Mark it as HAS CONTENT (result=false) if:
-- Contains actual issue description, even if brief
-- Has specific technical details, error messages, or implementation notes
-- Shows evidence that the template has been filled out with real information
-
-Respond ONLY in this JSON format:
-{
-  "result": false,
-  "reason": "Short explanation of why this is template-only or has content."
-}
-`.trim();
-
-    const result = await generateText({
-      model: openai(OPENAI_CONFIG.MODEL),
-      prompt,
-    });
-
-    const content = result.text?.trim();
-
-    if (!content) {
-      throw new Error('Empty LLM response');
-    }
-
-    const jsonMatch = content.match(/```(?:json)?([\s\S]*?)```/);
-    const raw = jsonMatch?.[1]?.trim() || content;
-
-    const parsed = JSON.parse(raw);
-
-    if (typeof parsed.result === 'boolean' && typeof parsed.reason === 'string') {
-      return parsed;
-    } else {
-      throw new Error('Missing or invalid fields in LLM response');
-    }
-  } catch (error) {
-    console.error('Error in LLM template detection:', error);
-    // Fallback to heuristic-based detection
-    const fallbackResult = fallbackTemplateDetection(body);
-    return {
-      result: fallbackResult,
-      reason: fallbackResult
-        ? 'Heuristic: Body appears to contain only template placeholders.'
-        : 'Heuristic: Body contains meaningful content beyond templates.',
-    };
-  }
-}
-
-/**
- * Detects unclear instructions using LLM
- */
-async function detectUnclearInstructions(title: string, body: string): Promise<{ result: boolean; reason: string }> {
-  if (!body || body.trim().length === 0) {
-    return {
-      result: false, // Empty body is handled by template detection
-      reason: 'Empty body is handled by template detection.',
-    };
-  }
-
-  try {
-    const prompt = `
-You are analyzing a GitHub Issue body to determine if it lacks clear, actionable instructions.
-
-Issue Title: ${title}
-Issue Body: ${body}
-
-Mark it as UNCLEAR (result=true) if:
-- Lacks specific, actionable steps or clear next actions
-- Vague or ambiguous description where expected outcome is unclear
-- Not enough context to understand what needs to be done
-- Very generic requests without specifics
-
-Mark it as CLEAR (result=false) if:
-- Contains concrete directives or specific tasks
-- References specific files, code areas, links, or resources
-- Has clear steps or requirements outlined
-- Provides sufficient context for someone to take action
-
-Respond ONLY in this JSON format:
-{
-  "result": false,
-  "reason": "Short explanation of why the instructions are clear or unclear."
-}
-`.trim();
-
-    const result = await generateText({
-      model: openai(OPENAI_CONFIG.MODEL),
-      prompt,
-    });
-
-    const content = result.text?.trim();
-
-    if (!content) {
-      throw new Error('Empty LLM response');
-    }
-
-    const jsonMatch = content.match(/```(?:json)?([\s\S]*?)```/);
-    const raw = jsonMatch?.[1]?.trim() || content;
-
-    const parsed = JSON.parse(raw);
-
-    if (typeof parsed.result === 'boolean' && typeof parsed.reason === 'string') {
-      return parsed;
-    } else {
-      throw new Error('Missing or invalid fields in LLM response');
-    }
-  } catch (error) {
-    console.error('Error in LLM unclear instructions detection:', error);
-    // Fallback to heuristic-based detection
-    const fallbackResult = fallbackUnclearInstructionsDetection(body);
-    return {
-      result: fallbackResult,
-      reason: fallbackResult
-        ? 'Heuristic: Body lacks minimum actionable clarity.'
-        : 'Heuristic: Body shows minimum actionable clarity.',
-    };
-  }
-}
-
-/**
- * Fallback heuristic for template-only detection.
- * Returns true if template-only; false if has content.
- */
-function fallbackTemplateDetection(body: string): boolean {
-  const normalized = body.trim().toLowerCase();
-
-  if (normalized.length === 0) return true;
-
-  const templatePhrases = [
-    'please describe the issue here',
-    'describe the problem',
-    'what did you expect to happen',
-    'what actually happened',
-    'please provide',
-    'fill in the details',
-    'add your description here',
-    'template content',
-    'placeholder text',
-    'rewrite the summary of the tasks',
-    'record the notes and requirements related to the order of merging',
-  ];
-
-  return templatePhrases.some((p) => normalized.includes(p));
-}
-
-/**
- * Fallback heuristic for unclear instructions detection.
- * Returns true if unclear; false if clear.
- */
-function fallbackUnclearInstructionsDetection(body: string): boolean {
-  const normalized = body.trim().toLowerCase();
-
-  if (normalized.length === 0) return false; // Empty body handled by template detection
-
-  // Signals of clarity (any makes it clear)
-  const hasList = /^[-*]\s+/m.test(body); // bullet points
-  const hasLink = /\bhttps?:\/\//i.test(body);
-  const hasFileRef = /\b\w+\.(ts|js|tsx|jsx|md|yml|yaml|json)\b/i.test(body) || /#\d+/.test(body);
-  const hasSection = /##\s*(purpose|spec|概要|詳細|タスク|task|steps|チェック|参考|memo|メモ|description)/i.test(body);
-  const actionVerbs = [
-    'fix','relax','update','add','remove','refactor','rewrite','implement',
-    'enable','disable','set','change','migrate','bump','link','document'
-  ];
-  const hasActionVerb = actionVerbs.some((v) => normalized.includes(v));
-
-  if (hasList || hasLink || hasFileRef || hasSection || hasActionVerb) {
-    // Considered clear enough even if short
-    return false;
-  }
-
-  // Vague vs technical signal
-  const vaguePhrases = [
-    'fix this',
-    'something is wrong',
-    "it doesn't work",
-    "there's an issue",
-    'problem with',
-    'needs to be fixed',
-    'broken',
-    'not working',
-    'help needed',
-  ];
-  const technicalTerms = [
-    'error',
-    'exception',
-    'stack trace',
-    'log',
-    'console',
-    'api',
-    'endpoint',
-    'database',
-    'query',
-    'function',
-    'method',
-    'class',
-    'component',
-    'file',
-    'line',
-    'code',
-    'unit test',
-    'jest',
-  ];
-
-  const vagueCount = vaguePhrases.filter((phrase) => normalized.includes(phrase)).length;
-  const technicalCount = technicalTerms.filter((term) => normalized.includes(term)).length;
-
-  // Extremely short with no clarity signals → unclear
-  if (normalized.length < 80) return true;
-
-  // Vague language + lack of technical/context → unclear (more lenient threshold)
-  if (vagueCount > 0 && technicalCount < 2 && normalized.length < 200) return true;
-
-  // Default: assume clear enough
-  return false;
-}
+// Also provide the named export for compatibility
+export { checkIssues } from './checkIssues/index';
